--- test/e2e/suites/allocate/applying_options/namespace_test.go
@@ -1,7 +1,0 @@
-/*
-Copyright 2025 NVIDIA CORPORATION
-SPDX-License-Identifier: Apache-2.0
-*/
-package applying_options
-
-var _ = DescribeNamespaceSpecs()
