--- test/e2e/suites/allocate/elastic/allocate_elastic_test.go
@@ -1,7 +1,0 @@
-/*
-Copyright 2025 NVIDIA CORPORATION
-SPDX-License-Identifier: Apache-2.0
-*/
-package elastic
-
-var _ = DescribeAllocateElasticSpecs()
