--- test/e2e/suites/allocate/predicates/taint_toleration_test.go
@@ -1,7 +1,0 @@
-/*
-Copyright 2025 NVIDIA CORPORATION
-SPDX-License-Identifier: Apache-2.0
-*/
-package predicates
-
-var _ = DescribeTaintTolerationSpecs()
