--- test/e2e/suites/allocate/quota/over_quota_test.go
@@ -1,7 +1,0 @@
-/*
-Copyright 2025 NVIDIA CORPORATION
-SPDX-License-Identifier: Apache-2.0
-*/
-package quota
-
-var _ = DescribeOverQuotaSpecs()
