+++ test/e2e/suites/allocate/quota/quota_suite_test.go
@@ -13,6 +13,10 @@ import (
 	. "github.com/onsi/gomega"
 )
 
+var _ = DescribeQuotaSpecs()
+var _ = DescribeExceedLimitSpecs()
+var _ = DescribeOverQuotaSpecs()
+
 func TestResourcesAllocation(t *testing.T) {
 	utils.SetLogger()
 	RegisterFailHandler(Fail)
