+++ test/e2e/suites/allocate/resources/resources_suite_test.go
@@ -13,6 +13,8 @@ import (
 	. "github.com/onsi/gomega"
 )
 
+var _ = DescribeResourcesSpecs()
+
 func TestResourcesAllocation(t *testing.T) {
 	utils.SetLogger()
 	RegisterFailHandler(Fail)
