--- test/e2e/suites/api/crds/podgroup/conditions_test.go
@@ -1,7 +1,0 @@
-/*
-Copyright 2025 NVIDIA CORPORATION
-SPDX-License-Identifier: Apache-2.0
-*/
-package podgroup
-
-var _ = DescribeConditionsSpecs()
