--- test/e2e/suites/api/events/events_test.go
@@ -1,7 +1,0 @@
-/*
-Copyright 2025 NVIDIA CORPORATION
-SPDX-License-Identifier: Apache-2.0
-*/
-package events
-
-var _ = DescribeEventsSpecs()
