+++ test/e2e/suites/consolidation/consolidation_suite_test.go
@@ -13,6 +13,8 @@ import (
 	. "github.com/onsi/gomega"
 )
 
+var _ = DescribeConsolidationSpecs()
+
 func TestConsolidation(t *testing.T) {
 	utils.SetLogger()
 	RegisterFailHandler(Fail)
