--- test/e2e/suites/preempt/preempt_pod_order_test.go
@@ -1,7 +1,0 @@
-/*
-Copyright 2025 NVIDIA CORPORATION
-SPDX-License-Identifier: Apache-2.0
-*/
-package preempt
-
-var _ = DescribePreemptPodOrderSpecs()
