--- test/e2e/suites/reclaim/hierarchy_level_fairness_test.go
@@ -1,7 +1,0 @@
-/*
-Copyright 2025 NVIDIA CORPORATION
-SPDX-License-Identifier: Apache-2.0
-*/
-package reclaim
-
-var _ = DescribeHierarchyLevelFairnessSpecs()
